+++ panda/src/dxgsg9/wdxGraphicsBuffer9.cxx
@@ -275,8 +275,8 @@ rebuild_bitplanes() {
   int bitplane_x = get_x_size();
   int bitplane_y = get_y_size();
   if (Texture::get_textures_power_2() != ATS_none) {
-    bitplane_x = Texture::up_to_power_2(bitplane_x);
-    bitplane_y = Texture::up_to_power_2(bitplane_y);
+    bitplane_x = (int)Texture::up_to_power_2(bitplane_x);
+    bitplane_y = (int)Texture::up_to_power_2(bitplane_y);
   }
 
   // Find the color and depth textures.  Either may be present, or neither.
